+++ src/Containers/VendorsList/VendorsHeader.stories.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Import } from '@styled-icons/boxicons-regular/Import';
 import { Add } from '@styled-icons/ionicons-sharp/Add';
-import { VendorsHeader, IVendorsHeaderProps } from '../../src';
-import { createStoryTitle } from '../Constants';
+import { VendorsHeader, IVendorsHeaderProps } from "../../index";
+import { createStoryTitle } from '../../Constants';
 import { Story, Meta } from '@storybook/react';
 
 export default {
