+++ src/Containers/VendorsList/VendorsHeader.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Button, ButtonProps } from '../../Inputs/Button';
+import { Button, ButtonProps } from '../../index';
 import { Heading } from '../../Text/Heading';
 import { MainInterface, ResponsiveInterface } from '../../Utils/BaseStyles';
 import { flex, media } from '../../Utils/Mixins';
